requirements.txt
flask
gunicorn
psycopg2-binary
Flask-SQLAlchemy
python-dotenv
Flask-Migrate



